+++ PIProjetpCards/Program.cs
@@ -17,7 +17,7 @@ namespace PIProjetpCards
         {
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
-            Application.Run(new LoginScreen());
+            Application.Run(new HelpScreen());
         }
     }
 }
