+++ PIProjetpCards/Screens/HelpScreen.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.Data;
@@ -10,12 +10,58 @@ using System.Windows.Forms;
 
 namespace PIProjetpCards.Screens
 {
-    public partial class HelpScreen : Form
+    public partial class HelpScreen: Form
     {
+
+        Dictionary<int, string> questoes = new Dictionary<int, string>()
+        {
+            { 0, "1 - O que são os FlashCards" },
+            { 1, "2 - Como utilizar os FlashCards?" },
+            { 2, "3 - Como alterar os dados?" },
+            { 3, "4 - Como criar notas?" },
+
+        };
+
+
+
+
+        Dictionary<int, string> respostas = new Dictionary<int, string>()
+        {
+            { 0, "Flashcards são pequenos cartões com perguntas e respostas sobre um conteúdo específico com a finalidade de avaliar a sua memória." +
+                      " Em outras palavras, trata-se de um card com perguntas de um lado, e respostas de outro, no qual é possível se autodesafiar. " },
+            { 1, "Usar flashcards é algo muito simples. Primeiro, você deve confeccioná-los à mão ou com a ajuda de sites e aplicativos." +
+                      " Lembre-se de investir em perguntas e respostas dinâmicas associadas aos conteúdos de estudo. A pergunta fica de um lado, " +
+                      "a resposta do outro. Inclusive, a dica é apostar em cores e letras diferentes para uma melhor memorização dos tópicos. !" },
+            { 2, "Para a alteração de dados, é necessário clicar no ícone de 'ajuda' e clicar no botão 'Dados' e fazer as alterações necessárias." },
+            { 3, "EM ATUALIZAÇÃO" }
+        };
+
+
         public HelpScreen()
         {
             InitializeComponent();
             this.FormClosing += applicationClose;
+        }
+
+        private void HelpScreen_Load(object sender, EventArgs e)
+        {
+            comboBox1.DataSource = new BindingSource(questoes, null);
+            comboBox1.DisplayMember = "Value";
+            comboBox1.ValueMember = "Key";
+        }
+
+        private void ComboBox1_SelectedIndexChanged(object sender, EventArgs e)
+        {
+            int opcaoSelecionada = comboBox1.SelectedIndex;
+
+            if (respostas.ContainsKey(opcaoSelecionada))
+            {
+                textBox1.Text = respostas[opcaoSelecionada];
+            }
+            else
+            {
+                textBox1.Text = "Opção inválida.";
+            }
         }
 
         private void applicationClose(object sender, FormClosingEventArgs e)
@@ -29,6 +75,13 @@ namespace PIProjetpCards.Screens
             {
                 Application.ExitThread();
             }
+        }
+
+        private void textBox1_TextChanged(object sender, EventArgs e)
+        {
+            textBox1.Multiline = true;
+            textBox1.ReadOnly = true;
+            textBox1.Enabled = false;
         }
     }
 }
