+++ PIProjetpCards/Screens/StatistcsScreen.resx
@@ -117,4 +117,15 @@
   <resheader name="writer">
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
+  <assembly alias="System.Drawing" name="System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a" />
+  <data name="btnBack.Image" type="System.Drawing.Bitmap, System.Drawing" mimetype="application/x-microsoft.net.object.bytearray.base64">
+    <value>
+        iVBORw0KGgoAAAANSUhEUgAAABgAAAAYCAYAAADgdz34AAAABGdBTUEAALGPC/xhBQAAAAlwSFlzAAAA
+        qAAAAKgBefSzxgAAABl0RVh0U29mdHdhcmUAd3d3Lmlua3NjYXBlLm9yZ5vuPBoAAACtSURBVEhL7ZTL
+        CQJBEEQfiNFoJmIgghkIexfdAPakKOhF/MSjYAZiBkrDCkPhabqHBfHBnKqodxhoyKcPNMADmGroxcYP
+        wKt9Vy140HF7lZZy6QF7Ga+1lEvx8d1//BsrGV9owcOo5LhxTsa3GkZwTwRDDSNIBQMNI7gkgo2GEYzl
+        k+daiMA+t6jELudJJEsteTHJsQtJ6Mkwflsy05IXldy0EIFJ1sATmGj44Q1wbVo/vGvTOwAAAABJRU5E
+        rkJggg==
+</value>
+  </data>
 </root>
