--- PIProjetpCards/StatisticsUser/StatisticsUser.cs
@@ -1,28 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Forms;
-using PIProjetpCards.Screens;
-
-namespace PIProjetpCards.StatisticsUser
-{
-    public partial class StatisticsUser : UserControl
-    {
-        public StatisticsUser()
-        {
-            InitializeComponent();
-        }
-
-        private void btnBack_Click(object sender, EventArgs e)
-        {
-            SettingsScreen settingsScreen = new SettingsScreen();
-            settingsScreen.Show();
-            this.ParentForm.Hide();
-        }
-    }
-}
